refactor(chat-ui): split session actions
