+++ apps/shared/OpenClawKit/Sources/OpenClawChatUI/ChatViewModel.swift
@@ -369,41 +369,6 @@ public final class OpenClawChatViewModel {
         Task { await self.performAbort() }
     }
 
-    public func refreshSessions(limit: Int? = nil) {
-        let context = self.currentSessionSnapshot()
-        Task { await self.fetchSessions(limit: limit, sessionSnapshot: context) }
-    }
-
-    public func startNewSession(worktree: Bool = false) async {
-        await self.performStartNewSession(worktree: worktree)
-    }
-
-    public func requestSessionReset() {
-        Task { await self.performReset() }
-    }
-
-    public func requestSessionCompact() {
-        Task { await self.performCompact() }
-    }
-
-    public func setSessionPinned(_ sessionKey: String, pinned: Bool) {
-        Task {
-            do {
-                try await self.transport.patchSession(
-                    key: sessionKey,
-                    label: nil,
-                    category: nil,
-                    pinned: pinned,
-                    archived: nil,
-                    unread: nil)
-            } catch {
-                self.errorText = error.localizedDescription
-                return
-            }
-            await self.fetchSessions(limit: nil, sessionSnapshot: self.currentSessionSnapshot())
-        }
-    }
-
     public func deleteSession(_ sessionKey: String) {
         Task {
             do {
@@ -1565,7 +1530,7 @@ public final class OpenClawChatViewModel {
         }
     }
 
-    private func fetchSessions(limit: Int?, sessionSnapshot: SessionSnapshot? = nil) async {
+    func fetchSessions(limit: Int?, sessionSnapshot: SessionSnapshot? = nil) async {
         do {
             let res = try await transport.listSessions(limit: limit, search: nil, archived: false)
             if let sessionSnapshot, !self.isCurrentSession(sessionSnapshot) { return }
@@ -1765,7 +1730,7 @@ public final class OpenClawChatViewModel {
         self.applySessionSwitch(to: sessionKey, intent: .externalSync)
     }
 
-    private func performStartNewSession(worktree: Bool) async {
+    func performStartNewSession(worktree: Bool) async {
         guard !self.blocksAttachmentOwnerChange else {
             self.errorText = String(
                 localized: "Remove attachments or wait for delivery to resolve before starting a new chat.")
@@ -1828,7 +1793,7 @@ public final class OpenClawChatViewModel {
             && nsError.localizedDescription == "sessions.create not supported by this transport"
     }
 
-    private func performReset() async {
+    func performReset() async {
         self.isLoading = true
         self.errorText = nil
 
@@ -1846,7 +1811,7 @@ public final class OpenClawChatViewModel {
         self.startBootstrap()
     }
 
-    private func performCompact() async {
+    func performCompact() async {
         guard !self.isCompacting else { return }
         guard !self.isSending, self.pendingRuns.isEmpty, !self.isAborting else {
             self.errorText = "Wait for the current response before compacting the session."
